+++ assets/system/shaders/galileo.fs.glsl
@@ -1,9 +1,16 @@
+/**
+ *  minisphere Galileo default fragment shader
+ *  provides the basic guarantees required by the Galileo specification
+ *  (c) 2015-2016 Fat Cerberus
+**/
+
 #ifdef GL_ES
 precision mediump float;
 #endif
 
 uniform sampler2D al_tex;
 uniform bool al_use_tex;
+
 varying vec4 varying_color;
 varying vec2 varying_texcoord;