+++ src/minisphere/screen.h
@@ -9,6 +9,7 @@ typedef struct screen screen_t;
 screen_t*        screen_new               (const char* title, image_t* icon, int x_size, int y_size, int frameskip, bool avoid_sleep);
 void             screen_free              (screen_t* obj);
 ALLEGRO_DISPLAY* screen_display           (const screen_t* obj);
+bool             screen_have_shaders      (const screen_t* screen);
 bool             screen_is_skipframe      (const screen_t* obj);
 rect_t           screen_get_clipping      (screen_t* obj);
 int              screen_get_frameskip     (const screen_t* obj);